+++ packages/resizable/resizable.js
@@ -86,8 +86,8 @@ Resizable.ContentWindow = class{
     /**
      * My additions
      */
-    console.log(div.id)
-    if (div.id == "main") { this.getDiv().style.overflow = "hidden"; } 
+    //console.log(div.id)
+    if (div.id == "main") { this.getDiv().style.overflow = "auto"; } 
     else { this.getDiv().style.overflow = "auto"; }
     //console.log(div.id);
     this.children = [];