+++ src/config.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import * as fs from "fs";
+//import * as fs from "fs";
 import * as path from "path";
 import { print, error } from "./src";
 import { ProposedExtensionAPI } from "./pythonApiTypes";
@@ -126,8 +126,8 @@ export async function getConfig(context:vscode.ExtensionContext):Promise<Config>
         // python module stuff
         config.pythonModuleName = "UC_Quantum_Lab";
         config.pythonModulePyPi = "UC-Quantum-tools";
-        config.minPythonModVer = "0.1.9";
-        config.minPythonVer = "3.6.0";
+        config.minPythonModVer = "0.1.10";
+        config.minPythonVer = "3.8.0";
 
         return config;
     } else {