+++ src/extension.ts
@@ -21,11 +21,8 @@ import { handleLegacy } from './handleLegacy';
  * @param config : configuration of the extension
  * @returns boolean indicating whether or not this function extension exceeded
  */
-async function init(config:Config) {
+async function init(config:Config, verbose:boolean) {
  	print("Running \"init\"");
-	
-	// verifying the selected python environment
-	await verifyPython(config);
 
 	// the config directory exists
 	if (!(fs.existsSync(config.configDir))) {
@@ -43,7 +40,8 @@ async function init(config:Config) {
 
 		// copies template layout file to be displayed in the viewer
 		try {
-			await fs.promises.copyFile(config.templateLayoutFile, config.layoutFile);
+			await vscode.workspace.fs.copy(vscode.Uri.file(config.templateLayoutFile), vscode.Uri.file(config.layoutFile));
+			//await fs.promises.copyFile(config.templateLayoutFile, config.layoutFile);
 		} catch ( e ) {
 			error(`while trying to copy template layout file to config dir: ${(e as Error).message}`);
 		}
@@ -60,13 +58,21 @@ async function init(config:Config) {
 			if (selection === config.yes) {
 				print("Making main file");
 				// copying template file to current directory
-				fs.copyFile(config.templatePythonFile, 
-							path.join(config.workspacePath, fname), 
-							(err) => {
-					if (err){
-						error(`Error copying "${config.templatePythonFile}" to "${fname}"`);
-					}
-				});
+				try{
+					vscode.workspace.fs.copy(
+						vscode.Uri.file(config.templatePythonFile), 
+						vscode.Uri.file(path.join(config.workspacePath, fname))
+					);
+				} catch ( e ) {
+					error(`Error copying "${config.templatePythonFile}" to "${fname}" with message: ${(e as Error).message}`);
+				}
+				// fs.copyFile(config.templatePythonFile, 
+				// 			path.join(config.workspacePath, fname), 
+				// 			(err) => {
+				// 	if (err){
+				// 		error(`Error copying "${config.templatePythonFile}" to "${fname}"`);
+				// 	}
+				// });
 				// opening the example file in the editor
 				let documet:vscode.TextDocument|undefined = await vscode.workspace.openTextDocument(path.join(config.workspacePath, fname));
 				if (documet === undefined) {
@@ -78,6 +84,8 @@ async function init(config:Config) {
 				}
 			}
 		}
+	} else if (verbose) {
+		info("Nothing to init, done");
 	}
 }
 
@@ -91,6 +99,7 @@ export async function activate(context: vscode.ExtensionContext) {
 	// adding the command to vscode
 	context.subscriptions.push(
 		vscode.commands.registerCommand("uc-quantum-lab.execute", async () => {
+			
 			print("--- executing ---");
 			try {
 				// loading the configuration from the ./config.ts
@@ -98,6 +107,18 @@ export async function activate(context: vscode.ExtensionContext) {
 				
 				// handles features from previous versions of this extension
 				await handleLegacy(config);
+
+				// verifying the selected python environment
+				await verifyPython(config);
+				// let watcher:vscode.FileSystemWatcher = vscode.workspace.createFileSystemWatcher(config.layoutFile, false, false, false);
+				// context.subscriptions.push(watcher);
+				// watcher.onDidCreate(() => {
+				// 	// updating the panel, when the layout file is updated
+				// 	UCQ.currentPanel?.update();
+				// });
+				// watcher.onDidDelete(() =>{
+				// 	watcher.dispose();
+				// });
 
 				// if the viewer panel is open and there is an active editor
 				if (UCQ.currentPanel && vscode.window.activeTextEditor) {
@@ -119,17 +140,38 @@ export async function activate(context: vscode.ExtensionContext) {
 					} else {
 						print("executing in termial");
 						// executing the python file in the terminal with the python extension
-						vscode.commands.executeCommand("python.execInTerminal");
+						// vscode.commands.executeCommand("python.execInTerminal");
+						// if here, then the file is a python file
+						print("saving active document");
+						await vscode.window.activeTextEditor.document.save();
+						
+						print("executing in termial");
+						// if there is an active terminal in editor
+						if (vscode.window.activeTerminal) {
+							print("Sending to active terminal");
+							// making sure the user can see the terminal
+							vscode.window.activeTerminal.show(true);
+							// sending the python command to active terminal to execute the active python file
+							vscode.window.activeTerminal.sendText(`${config.userConfig.python} ${vscode.window.activeTextEditor.document.fileName}`);
+						} else {
+							// if here, then there was no active terminal so one is made
+							print("creating terminal and sending to it");
+							// creating terminal in vscode
+							let term = vscode.window.createTerminal();
+							// sending the python command to active terminal to execute the active python file
+							term.sendText(`${config.userConfig.python} ${vscode.window.activeTextEditor.document.fileName}`);
+						}
 						// waiting for the layout.json file to be updated
 						let watcher:vscode.FileSystemWatcher = vscode.workspace.createFileSystemWatcher(config.layoutFile, false, false, false);
 						watcher.onDidChange(() => {
 							// updating the panel, when the layout file is updated
 							UCQ.currentPanel?.update();
+							watcher.dispose();	
 						});
 					}
 				} else {
 					// if nothing is opening, first running init to make sure everything is setup correctly
-					await init(config);
+					await init(config, false);
 					print("Creating Window");
 					// creating window
 					UCQ.createOrShow(config);
@@ -147,7 +189,7 @@ export async function activate(context: vscode.ExtensionContext) {
 				let config:Config = await getConfig(context);
 				
 				// initing the current directory
-				await init(config);
+				await init(config, true);
 			// functions handle their own errors so do not need to do anything here
 			} catch ( e ) {}
 		})
@@ -167,8 +209,10 @@ export async function activate(context: vscode.ExtensionContext) {
 					// if they agreed to the previous message
 					if (choice === config.yes) {
 						// removing the config directory and catching errors
-						try { fs.rmSync(config.configDir, { recursive: true, force: true }); }
-						catch ( e ) { error(`error encountered when deleting config directory, with message: ${e}`); }
+						try { 
+							await vscode.workspace.fs.delete(vscode.Uri.file(config.configDir), {recursive:true});
+							//fs.rmSync(config.configDir, { recursive: true, force: true }); 
+						} catch ( e ) { error(`error encountered when deleting config directory, with message: ${e}`); }
 
 					} else if (choice === undefined) {
 						// the user can not choose nothing
@@ -177,7 +221,7 @@ export async function activate(context: vscode.ExtensionContext) {
 					}
 				}
 				// initing the current directory, do not to wrap the function call in an "if" because the function will handle its own errors
-				await init(config);
+				await init(config, true);
 			// functions handle their own errors so do not need to do anything here
 			} catch ( e ) {}
 		})